+++ apps/client/src/app/core/database/custom-links/custom-link.ts
@@ -24,14 +24,14 @@ export class CustomLink extends DataModel {
     return type;
   }
 
-  getUrl(): string {
+  getUrl = () => {
     // If we're inside Electron, create a direct Teamcraft link.
     if (navigator.userAgent.toLowerCase().indexOf('electron/') > -1) {
       return `https://ffxivteamcraft.com/link/${encodeURI(this.authorNickname)}/${encodeURI(this.uri)}`;
     } else {
       return `${window.location.protocol}//${window.location.host}/link/${encodeURI(this.authorNickname)}/${encodeURI(this.uri)}`;
     }
-  }
+  };
 
   getEntityId(): string {
     const exploded = this.redirectTo.split('/');