+++ apps/client/src/app/modules/list/list-details-panel/list-details-panel.component.ts
@@ -67,13 +67,12 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
   finalItems = false;
 
   @Input()
-  collapsed = false;
-
-  @Input()
   overlay = false;
 
   @Input()
   largeList = false;
+
+  collapsed = false;
 
   progression: number;
 
@@ -154,6 +153,8 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
 
   ngOnInit(): void {
     if (this.displayRow && this.displayRow.collapsedByDefault) {
+      this.collapsed = true;
+    } else if (this.displayRow && this.displayRow.rows.length > 25) {
       this.collapsed = true;
     }
   }
@@ -457,7 +458,7 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
     });
   }
 
-  public getTextExport(): string {
+  public getTextExport = () => {
     let rows: ListRow[];
     if (this.tiers) {
       rows = this.tiers.reduce((res, tier) => {
@@ -469,11 +470,7 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
     return rows.reduce((exportString, row) => {
       return exportString + `${row.amount}x ${this.i18nTools.getName(this.l12n.getItem(row.id))}\n`;
     }, `${this.displayRow.title} :\n`);
-  }
-
-  textCopied(): void {
-    this.message.success(this.translate.instant('LIST.Copied_as_text'));
-  }
+  };
 
   trackByItem(index: number, item: ListRow): number {
     return item.id;