+++ apps/client/src/app/pages/recipe-finder/recipe-finder.module.ts
@@ -4,7 +4,6 @@ import { RecipeFinderComponent } from './recipe-finder/recipe-finder.component';
 import { FormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
 import { NgZorroAntdModule } from 'ng-zorro-antd';
-import { ClipboardModule } from 'ngx-clipboard';
 import { PipesModule } from '../../pipes/pipes.module';
 import { CoreModule } from '../../core/core.module';
 import { ListModule } from '../../modules/list/list.module';
@@ -33,13 +32,11 @@ const routes: Routes = [
     FormsModule,
     TranslateModule,
     NgZorroAntdModule,
-    ClipboardModule,
     PipesModule,
     CoreModule,
     ListModule,
     FlexLayoutModule,
     FullpageMessageModule,
-    ClipboardModule,
 
     RouterModule.forChild(routes),
     ItemIconModule,