+++ apps/client/src/app/pages/search/search.module.ts
@@ -20,7 +20,7 @@ import { SearchIntroComponent } from './search-intro/search-intro.component';
 import { ItemIconModule } from '../../modules/item-icon/item-icon.module';
 import { MaintenanceGuard } from '../maintenance/maintenance.guard';
 import { VersionLockGuard } from '../version-lock/version-lock.guard';
-import { ClipboardModule } from 'ngx-clipboard';
+
 import { SearchResultComponent } from './search-result/search-result.component';
 
 const routes: Routes = [
@@ -37,7 +37,7 @@ const routes: Routes = [
     FormsModule,
     ReactiveFormsModule,
     FlexLayoutModule,
-    ClipboardModule,
+
 
     CoreModule,
     ListModule,